@@ -6,81 +6,80 @@ using Backend.Fx.Exceptions;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Http;
 
-namespace Backend.Fx.AspNetCore.ErrorHandling
+namespace Backend.Fx.AspNetCore.ErrorHandling;
+
+public abstract class ErrorHandlingMiddleware
 {
-    public abstract class ErrorHandlingMiddleware
+    private readonly RequestDelegate _next;
+
+    /// <summary>
+    ///     This constructor is being called by the framework DI container
+    /// </summary>
+    [UsedImplicitly]
+    protected ErrorHandlingMiddleware(RequestDelegate next)
     {
-        private readonly RequestDelegate _next;
+        _next = next;
+    }
 
-        /// <summary>
-        ///     This constructor is being called by the framework DI container
-        /// </summary>
-        [UsedImplicitly]
-        protected ErrorHandlingMiddleware(RequestDelegate next)
+    /// <summary>
+    ///     This method is being called by the previous middleware in the HTTP pipeline
+    /// </summary>
+    [UsedImplicitly]
+    public async Task Invoke(HttpContext context)
+    {
+        if (await ShouldHandle(context))
         {
-            _next = next;
-        }
-
-        /// <summary>
-        ///     This method is being called by the previous middleware in the HTTP pipeline
-        /// </summary>
-        [UsedImplicitly]
-        public async Task Invoke(HttpContext context)
-        {
-            if (await ShouldHandle(context))
-            {
-                try
-                {
-                    await _next.Invoke(context);
-                }
-                catch (TooManyRequestsException tmrex)
-                {
-                    if (tmrex.RetryAfter > 0)
-                    {
-                        context.Response.Headers.Add("Retry-After", tmrex.RetryAfter.ToString(CultureInfo.InvariantCulture));
-                    }
-
-                    await HandleClientError(context, 429, "TooManyRequests", tmrex);
-                }
-                catch (UnprocessableException uex)
-                {
-                    await HandleClientError(context, 422, "Unprocessable", uex);
-                }
-                catch (NotFoundException nfex)
-                {
-                    await HandleClientError(context, (int) HttpStatusCode.NotFound, HttpStatusCode.NotFound.ToString(), nfex);
-                }
-                catch (ConflictedException confex)
-                {
-                    await HandleClientError(context, (int) HttpStatusCode.Conflict, HttpStatusCode.Conflict.ToString(), confex);
-                }
-                catch (ForbiddenException uex)
-                {
-                    await HandleClientError(context, (int) HttpStatusCode.Forbidden, HttpStatusCode.Forbidden.ToString(), uex);
-                }
-                catch (UnauthorizedException uex)
-                {
-                    await HandleClientError(context, (int) HttpStatusCode.Unauthorized, HttpStatusCode.Unauthorized.ToString(), uex);
-                }
-                catch (ClientException cex)
-                {
-                    await HandleClientError(context, (int) HttpStatusCode.BadRequest, HttpStatusCode.BadRequest.ToString(), cex);
-                }
-                catch (Exception ex)
-                {
-                    await HandleServerError(context, ex);
-                }
-            }
-            else
+            try
             {
                 await _next.Invoke(context);
             }
+            catch (TooManyRequestsException tooManyRequestsException)
+            {
+                if (tooManyRequestsException.RetryAfter > 0)
+                {
+                    context.Response.Headers.Add("Retry-After", tooManyRequestsException.RetryAfter.ToString(CultureInfo.InvariantCulture));
+                }
+
+                await HandleClientError(context, 429, "TooManyRequests", tooManyRequestsException);
+            }
+            catch (UnprocessableException uex)
+            {
+                await HandleClientError(context, 422, "Unprocessable", uex);
+            }
+            catch (NotFoundException notFoundException)
+            {
+                await HandleClientError(context, (int) HttpStatusCode.NotFound, HttpStatusCode.NotFound.ToString(), notFoundException);
+            }
+            catch (ConflictedException conflictedException)
+            {
+                await HandleClientError(context, (int) HttpStatusCode.Conflict, HttpStatusCode.Conflict.ToString(), conflictedException);
+            }
+            catch (ForbiddenException forbiddenException)
+            {
+                await HandleClientError(context, (int) HttpStatusCode.Forbidden, HttpStatusCode.Forbidden.ToString(), forbiddenException);
+            }
+            catch (UnauthorizedException unauthorizedException)
+            {
+                await HandleClientError(context, (int) HttpStatusCode.Unauthorized, HttpStatusCode.Unauthorized.ToString(), unauthorizedException);
+            }
+            catch (ClientException clientException)
+            {
+                await HandleClientError(context, (int) HttpStatusCode.BadRequest, HttpStatusCode.BadRequest.ToString(), clientException);
+            }
+            catch (Exception exception)
+            {
+                await HandleServerError(context, exception);
+            }
         }
-
-        protected abstract Task<bool> ShouldHandle(HttpContext context);
-
-        protected abstract Task HandleClientError(HttpContext context, int httpStatusCode, string message, ClientException exception);
-
-        protected abstract Task HandleServerError(HttpContext context, Exception exception);
+        else
+        {
+            await _next.Invoke(context);
+        }
     }
+
+    protected abstract Task<bool> ShouldHandle(HttpContext context);
+
+    protected abstract Task HandleClientError(HttpContext context, int httpStatusCode, string message, ClientException exception);
+
+    protected abstract Task HandleServerError(HttpContext context, Exception exception);
 }
