@@ -1,40 +1,32 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
 using Backend.Fx.Exceptions;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 
-namespace Backend.Fx.AspNetCore.Mvc.Validation
+namespace Backend.Fx.AspNetCore.Mvc.Validation;
+
+public static class ModelStateEx
 {
-    public static class ModelStateEx
+    public static Errors ToErrorsDictionary(this ModelStateDictionary modelState)
     {
-        public static string ToDebugString(this ModelStateDictionary modelState)
+        var dictionary = new Dictionary<string, string[]>();
+        
+        foreach (var keyValuePair in modelState)
         {
-            var modelErrorMessages = modelState
-                                     .Where(kvp => kvp.Value.ValidationState == ModelValidationState.Invalid)
-                                     .Select(kvp => $"{kvp.Key}: {string.Join(", ", kvp.Value.Errors.Select(err => err.ErrorMessage))}");
-
-            return string.Join(System.Environment.NewLine, modelErrorMessages);
+            dictionary.Add(keyValuePair.Key, keyValuePair.Value.Errors.Select(err => err.ErrorMessage).ToArray());    
         }
 
-        public static Errors ToErrorsDictionary(this ModelStateDictionary modelState)
+        return new Errors(dictionary);
+    }
+
+
+    public static void Add(this ModelStateDictionary modelState, Errors errors)
+    {
+        foreach (var (key, value) in errors)
         {
-            var errors = new Errors();
-            foreach (var keyValuePair in modelState)
+            foreach (var errorMessage in value)
             {
-                errors.Add(keyValuePair.Key, keyValuePair.Value.Errors.Select(err => err.ErrorMessage));
-            }
-
-            return errors;
-        }
-
-
-        public static void Add(this ModelStateDictionary modelState, Errors errors)
-        {
-            foreach (var (key, value) in errors)
-            {
-                foreach (var errorMessage in value)
-                {
-                    modelState.AddModelError(key, errorMessage);
-                }
+                modelState.AddModelError(key, errorMessage);
             }
         }
     }
