@@ -1,32 +1,34 @@
-﻿using Backend.Fx.Exceptions;
+﻿using Backend.Fx.AspNetCore.ErrorHandling;
+using Backend.Fx.Exceptions;
+using JetBrains.Annotations;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace Backend.Fx.AspNetCore.Mvc.Validation
+namespace Backend.Fx.AspNetCore.Mvc.Validation;
+
+/// <summary>
+/// Returns HTTP 400 "Bad Request" when model validation failed. In addition, the bad model state is converted
+/// into an <see cref="ErrorResponse"/>s that is returned as JSON body, if the request
+/// stated that it accepts JSON as response type content.
+/// </summary>
+[PublicAPI]
+public class ReturnModelStateAsJsonModelValidationFilter : ModelValidationFilter
 {
-    /// <summary>
-    /// Returns HTTP 400 "Bad Request" when model validation failed. In addition, the bad model state is converted into an instance of
-    /// <see cref="Errors"/> gets serialized to the body as JSON.
-    /// </summary>
-    public class ReturnModelStateAsJsonModelValidationFilter : ModelValidationFilter
+    public override void OnActionExecuting(ActionExecutingContext context)
     {
-        public override void OnActionExecuting(ActionExecutingContext context)
-        {
-            if (!context.ModelState.IsValid && AcceptsJson(context))
-            {
-                Errors errors = context.ModelState.ToErrorsDictionary();
-                LogErrors(context, context.Controller.ToString(), errors);
-                context.Result = CreateResult(errors);
-            }
-        }
+        if (context.ModelState.IsValid || !AcceptsJson(context)) return;
+            
+        Errors errors = context.ModelState.ToErrorsDictionary();
+        LogErrors(context, context.Controller.ToString(), errors);
+        context.Result = CreateResult(errors);
+    }
 
-        protected virtual IActionResult CreateResult(Errors errors)
-        {
-            return new BadRequestObjectResult(errors);
-        }
+    protected virtual IActionResult CreateResult(Errors errors)
+    {
+        return new JsonResult(new ErrorResponse(errors));
+    }
 
-        public override void OnActionExecuted(ActionExecutedContext context)
-        {
-        }
+    public override void OnActionExecuted(ActionExecutedContext context)
+    {
     }
 }
