@@ -1,88 +1,96 @@
-using System;
+using System.Collections.Generic;
 using System.Diagnostics;
-using System.Threading;
-using Backend.Fx.Patterns.DependencyInjection;
-using Backend.Fx.Patterns.EventAggregation.Integration;
-using Backend.Fx.Tests;
+using System.Linq;
+using System.Threading.Tasks;
+using Backend.Fx.Features.MessageBus;
+using Backend.Fx.Logging;
+using Backend.Fx.SimpleInjectorDependencyInjection;
+using Backend.Fx.TestUtil;
 using FakeItEasy;
-using RabbitMQ.Client;
+using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Backend.Fx.RabbitMq.Tests
+namespace Backend.Fx.RabbitMq.Tests;
+
+public class TheRabbitMqMessageBus : TestWithLogging
 {
-    public class TheRabbitMqMessageBus : TestWithLogging
+    private readonly IBackendFxApplication _sender;
+    private readonly IBackendFxApplication _recipient;
+    private readonly IExceptionLogger _exceptionLogger = A.Fake<IExceptionLogger>();
+
+    // requires rabbitmq running on localhost:
+    // docker run -p 5672:5672 -e RABBITMQ_DEFAULT_USER=test -e RABBITMQ_DEFAULT_PASS=password rabbitmq
+    private readonly RabbitMqOptions _options = new()
     {
-        private readonly ManualResetEvent _received = new ManualResetEvent(false);
-        private readonly BackendFxApplicationInvoker _senderInvoker;
-        private readonly BackendFxApplicationInvoker _receiverInvoker;
+        Hostname = "localhost",
+        Username = "test",
+        Password = "password",
+        ExchangeName = "test-exchange",
+        ReceiveQueueName = "test-queue",
+        RetryCount = 3,
+    };
 
-        public TheRabbitMqMessageBus(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
+    public TheRabbitMqMessageBus(ITestOutputHelper output) : base(output)
+    {
+        _sender = new BackendFxApplication(new SimpleInjectorCompositionRoot(), _exceptionLogger);
+        _sender.EnableFeature(new MessageBusFeature(new RabbitMqMessageBus(_options)));
+        
+        _recipient = new BackendFxApplication(new SimpleInjectorCompositionRoot(), _exceptionLogger, GetType().Assembly);
+        _recipient.EnableFeature(new MessageBusFeature(new RabbitMqMessageBus(_options)));
+    }
+
+    [Fact]
+    public async Task InjectsMessageBusScope()
+    {
+        await _sender.BootAsync();
+        
+        await _sender.Invoker.InvokeAsync(sp =>
         {
-            var fakeSenderApplication = A.Fake<IBackendFxApplication>();
-            _senderInvoker = new BackendFxApplicationInvoker(fakeSenderApplication.CompositionRoot);
-
-
-            var fakeReceiverApplication = A.Fake<IBackendFxApplication>();
-            _receiverInvoker = new BackendFxApplicationInvoker(fakeReceiverApplication.CompositionRoot);
-            var fakeScope = A.Fake<IInjectionScope>();
-            var fakeInstanceProvider = A.Fake<IInstanceProvider>();
-            A.CallTo(() => fakeReceiverApplication.CompositionRoot.BeginScope()).Returns(fakeScope);
-            A.CallTo(() => fakeScope.InstanceProvider).Returns(fakeInstanceProvider);
-            A.CallTo(() => fakeInstanceProvider.GetInstance(A<Type>.That.IsEqualTo(typeof(TestIntegrationEventHandler))))
-             .Returns(new TestIntegrationEventHandler(_received));
-        }
-
-        //[Fact]
-        public void CanBeUsedWithBackendFxApplication()
+            var messageBusScope = sp.GetRequiredService<IMessageBusScope>();
+            Assert.NotNull(messageBusScope);
+            return Task.CompletedTask;
+        });
+    }
+    
+    [Fact]
+    public async Task InjectsHandlersAsCollections()
+    {
+        await _recipient.BootAsync();
+        await _recipient.Invoker.InvokeAsync(sp =>
         {
-            IMessageBus sender = new RabbitMqMessageBus(new ConnectionFactory
-            {
-                HostName = "localhost",
-                UserName = "anicors",
-                Password = "R4bb!tMQ"
-            }, 5, "unittest", "testSender");
-            sender.ProvideInvoker(_senderInvoker);
-            sender.Connect();
+            var handlers = sp.GetRequiredService<IEnumerable<IIntegrationEventHandler<DummyIntegrationEvent>>>().ToArray();
+            Assert.NotNull(handlers);
+            Assert.IsType<DummyIntegrationEventHandler>(Assert.Single(handlers));
+            return Task.CompletedTask;
+        });
+    }
 
-            var receiver = new RabbitMqMessageBus(new ConnectionFactory
-            {
-                HostName = "localhost",
-                UserName = "anicors",
-                Password = "R4bb!tMQ"
-            }, 5, "unittest", "testReceiver");
-
-            receiver.ProvideInvoker(_receiverInvoker);
-            receiver.Connect();
-            receiver.Subscribe<TestIntegrationEventHandler, TestIntegrationEvent>();
-
-            sender.Publish(new TestIntegrationEvent(1));
-            Assert.True(_received.WaitOne(Debugger.IsAttached ? int.MaxValue : 5000));
-        }
-
-        public class TestIntegrationEventHandler : IIntegrationMessageHandler<TestIntegrationEvent>
+    //todo fixme [Fact]
+    public async Task PublishesAndHandlesEvents()
+    {
+        // we have two applications
+        await _sender.BootAsync();
+        await _recipient.BootAsync();
+        
+        // we send an integration event on the first app
+        var dummyIntegrationEvent = new DummyIntegrationEvent();
+        await _sender.Invoker.InvokeAsync(sp =>
         {
-            private readonly ManualResetEvent _received;
+            var messageBusScope = sp.GetRequiredService<IMessageBusScope>();
+            messageBusScope.Publish(dummyIntegrationEvent);
+            return Task.CompletedTask;
+        });
 
-            public TestIntegrationEventHandler(ManualResetEvent received)
-            {
-                _received = received;
-            }
+        Assert.True(DummyIntegrationEventHandler.Called.WaitOne(Debugger.IsAttached ? int.MaxValue : 3000));
+    }
 
-            public void Handle(TestIntegrationEvent eventData)
-            {
-                _received.Set();
-            }
-        }
-
-        public class TestIntegrationEvent : IntegrationEvent
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
         {
-            public TestIntegrationEvent(int sequencenumber) : base()
-            {
-                Sequencenumber = sequencenumber;
-            }
-
-            public int Sequencenumber { get; }
+            _sender.Dispose();
+            _recipient.Dispose();
         }
     }
 }
