@@ -1,11 +1,13 @@
 ﻿using System;
+using JetBrains.Annotations;
 
 namespace Backend.Fx.Exceptions
 {
+    [PublicAPI]
     public class ForbiddenException : ClientException
     {
         public ForbiddenException()
-            : base("Unauthorized")
+            : base("Forbidden")
         {
         }
 
