@@ -2,9 +2,11 @@
 using System.Diagnostics;
 using System.Linq.Expressions;
 using System.Reflection;
+using JetBrains.Annotations;
 
 namespace Backend.Fx.Hacking
 {
+    [PublicAPI]
     public static class PrivateSetterCaller
     {
         public static void SetPrivate<T, TValue>(this T instance, Expression<Func<T, TValue>> propertyExpression, TValue value)
@@ -14,10 +16,10 @@ namespace Backend.Fx.Hacking
 
         private static string GetName<T, TValue>(Expression<Func<T, TValue>> exp)
         {
-            if (!(exp.Body is MemberExpression body))
+            if (exp.Body is not MemberExpression body)
             {
-                var ubody = (UnaryExpression) exp.Body;
-                body = ubody.Operand as MemberExpression;
+                var unaryExpression = (UnaryExpression) exp.Body;
+                body = unaryExpression.Operand as MemberExpression;
             }
 
             Debug.Assert(body != null, "body != null");
