@@ -1,13 +1,15 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 
 namespace Backend.Fx.Logging
 {
+    [PublicAPI]
     public class ExceptionLoggers : ICollection<IExceptionLogger>, IExceptionLogger
     {
-        private static readonly Microsoft.Extensions.Logging.ILogger Logger = Log.Create<ExceptionLoggers>();
+        private readonly ILogger _logger = Log.Create<ExceptionLoggers>();
         private readonly ICollection<IExceptionLogger> _collectionImplementation = new List<IExceptionLogger>();
 
         public ExceptionLoggers()
@@ -32,7 +34,7 @@ namespace Backend.Fx.Logging
                 }
                 catch (Exception ex2)
                 {
-                    Logger.LogError(ex,
+                    _logger.LogError(ex,
                         "{ExceptionLoggerTypeName} failed to log the {ExceptionTypeName} with message {ExceptionMessage}",
                         exceptionLogger.GetType().Name,
                         ex2.GetType(),
