@@ -1,22 +1,27 @@
-using Backend.Fx.AspNetCore.Tests.SampleApp.Runtime;
+using Backend.Fx.AspNetCore.Bootstrapping;
 using Backend.Fx.Environment.MultiTenancy;
 using Backend.Fx.InMemoryPersistence;
 using Backend.Fx.Logging;
-using Backend.Fx.Patterns.DependencyInjection;
-using Backend.Fx.Patterns.EventAggregation.Integration;
+using Backend.Fx.SimpleInjectorDependencyInjection;
 
 namespace Backend.Fx.AspNetCore.Tests.SampleApp
 {
     public class SampleApplicationHostedService : BackendFxApplicationHostedService
     {
-        public  ITenantService TenantService { get; }
-        public override IBackendFxApplication Application { get; }
-
         public SampleApplicationHostedService(IExceptionLogger exceptionLogger)
         {
-            IMessageBus messageBus = new InMemoryMessageBus();
-            TenantService = new TenantService(messageBus, new InMemoryTenantRepository());
-            Application = new SampleApplication(TenantService.TenantIdProvider, exceptionLogger);
+            IBackendFxApplication application = new BackendFxApplication(
+                new SimpleInjectorCompositionRoot(),
+                exceptionLogger,
+                typeof(SampleApplicationHostedService).Assembly);
+
+            application = new AspNetCoreApplication(application);
+
+            Application = application;
         }
+        
+        public  ITenantService TenantService { get; } = new TenantService(new InMemoryTenantRepository());
+        
+        public override IBackendFxApplication Application { get; }
     }
 }
