@@ -2,7 +2,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Backend.Fx.AspNetCore.Tests.SampleApp.Domain;
-using Backend.Fx.Tests;
+using Backend.Fx.TestUtil;
 using Newtonsoft.Json;
 using Xunit;
 using Xunit.Abstractions;
@@ -16,7 +16,7 @@ namespace Backend.Fx.AspNetCore.Tests
         
         public TheBackendFxMvcApplication(ITestOutputHelper output) : base(output)
         {
-            _factory = new SampleAppWebApplicationFactory(base.Logger);
+            _factory = new SampleAppWebApplicationFactory(Logger);
         }
         
         [Fact]
