@@ -1,5 +1,7 @@
-﻿using Backend.Fx.Exceptions;
-using Backend.Fx.Tests.BuildingBlocks;
+﻿using Backend.Fx.Domain;
+using Backend.Fx.Exceptions;
+using Backend.Fx.TestUtil;
+using JetBrains.Annotations;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -7,16 +9,29 @@ namespace Backend.Fx.Tests.Exceptions
 {
     public class TheNotFoundException : TestWithLogging
     {
+        public TheNotFoundException(ITestOutputHelper output) : base(output)
+        { }
+
+        [Fact]
+        public void CanBeThrownWithoutAnyParameters()
+        {
+            var exception = new NotFoundException();
+            Assert.Null(exception.EntityName);
+        }
+        
         [Fact]
         public void FillsNameAndIdProperties()
         {
-            var exception = new NotFoundException<TheAggregateRoot.TestAggregateRoot>(4711);
-            Assert.Equal("TestAggregateRoot", exception.EntityName);
+            var exception = new NotFoundException<SomeEntity>(4711);
+            Assert.Equal("SomeEntity", exception.EntityName);
             Assert.Equal(4711, exception.Id);
         }
 
-        public TheNotFoundException(ITestOutputHelper output) : base(output)
+
+        [UsedImplicitly]
+        private class SomeEntity : IAggregateRoot<int>
         {
+            public int Id { get; }
         }
     }
 }
